feat(interchain-indexer): Stats zero-count chain visibility setting (#1662)
* Added `StatsSettings`. Introducing `include_zero_chains`.

* Added ENV validation step into the implementation-plan workflow

* Update memory bank

* Code review issues

* Fix issues

+++ interchain-indexer/interchain-indexer-logic/src/chain_info/service.rs
@@ -180,7 +180,7 @@ impl ChainInfoService {
             let has_name = has_valid_name(&chain);
             let normalized = normalize_chain(chain);
             if has_name {
-                cache.insert(normalized.id as i64, normalized);
+                cache.insert(normalized.id, normalized);
             }
         }
